+++ src/TreeView/TreeView.Component.js
@@ -1,204 +1,204 @@
 import path from 'path';
 import React from 'react';
+import { TreeView } from '../';
 import { Description, DocsText, DocsTile, Header, Import, Properties, Separator } from '../_playground';
-import { Tree, TreeBranch, TreeCol, TreeHead, TreeItem, TreeRow, TreeView } from '../';
 
 export const TreeViewComponent = () => {
     const simpleTreeViewCode = `
 <TreeView>
-    <Tree>
-        <TreeItem>
-            <TreeRow>
-                <TreeCol>Row 1</TreeCol>
-            </TreeRow>
-        </TreeItem>
-        <TreeItem>
-            <TreeRow>
-                <TreeCol>Row 2</TreeCol>
-            </TreeRow>
-            <TreeBranch>
-                <TreeItem>
-                    <TreeRow>
-                        <TreeCol>Child 1</TreeCol>
-                    </TreeRow>
-                </TreeItem>
-                <TreeItem>
-                    <TreeRow>
-                        <TreeCol>Child 2</TreeCol>
-                    </TreeRow>
-                </TreeItem>
-            </TreeBranch>
-        </TreeItem>
-        <TreeItem>
-            <TreeRow>
-                <TreeCol>Row 3</TreeCol>
-            </TreeRow>
-            <TreeBranch>
-                <TreeItem>
-                    <TreeRow>
-                        <TreeCol>Child 1</TreeCol>
-                    </TreeRow>
-                    <TreeBranch>
-                        <TreeItem>
-                            <TreeRow>
-                                <TreeCol>Grandchild 1</TreeCol>
-                            </TreeRow>
-                        </TreeItem>
-                        <TreeItem>
-                            <TreeRow>
-                                <TreeCol>Grandchild 2</TreeCol>
-                            </TreeRow>
-                        </TreeItem>
-                    </TreeBranch>
-                </TreeItem>
-            </TreeBranch>
-        </TreeItem>
-    </Tree>
+    <TreeView.Tree>
+        <TreeView.Item>
+            <TreeView.Row>
+                <TreeView.Col>Row 1</TreeView.Col>
+            </TreeView.Row>
+        </TreeView.Item>
+        <TreeView.Item>
+            <TreeView.Row>
+                <TreeView.Col>Row 2</TreeView.Col>
+            </TreeView.Row>
+            <TreeView.Branch>
+                <TreeView.Item>
+                    <TreeView.Row>
+                        <TreeView.Col>Child 1</TreeView.Col>
+                    </TreeView.Row>
+                </TreeView.Item>
+                <TreeView.Item>
+                    <TreeView.Row>
+                        <TreeView.Col>Child 2</TreeView.Col>
+                    </TreeView.Row>
+                </TreeView.Item>
+            </TreeView.Branch>
+        </TreeView.Item>
+        <TreeView.Item>
+            <TreeView.Row>
+                <TreeView.Col>Row 3</TreeView.Col>
+            </TreeView.Row>
+            <TreeView.Branch>
+                <TreeView.Item>
+                    <TreeView.Row>
+                        <TreeView.Col>Child 1</TreeView.Col>
+                    </TreeView.Row>
+                    <TreeView.Branch>
+                        <TreeView.Item>
+                            <TreeView.Row>
+                                <TreeView.Col>Grandchild 1</TreeView.Col>
+                            </TreeView.Row>
+                        </TreeView.Item>
+                        <TreeView.Item>
+                            <TreeView.Row>
+                                <TreeView.Col>Grandchild 2</TreeView.Col>
+                            </TreeView.Row>
+                        </TreeView.Item>
+                    </TreeView.Branch>
+                </TreeView.Item>
+            </TreeView.Branch>
+        </TreeView.Item>
+    </TreeView.Tree>
 </TreeView>`;
 
     const treeViewCode = `
 <TreeView>
-    <TreeHead>
-        <TreeCol>Column Header 1</TreeCol>
-        <TreeCol>Column Header 2</TreeCol>
-        <TreeCol>Column Header 3</TreeCol>
-        <TreeCol>Column Header 4</TreeCol>
-    </TreeHead>
-    <Tree>
-        <TreeItem>
-            <TreeRow>
-                <TreeCol>Row 1</TreeCol>
-                <TreeCol>Data Col 2</TreeCol>
-                <TreeCol>Data Col 3</TreeCol>
-                <TreeCol>Data Col 4</TreeCol>
-            </TreeRow>
-            <TreeBranch>
-                <TreeItem>
-                    <TreeRow>
-                        <TreeCol>Child 1</TreeCol>
-                        <TreeCol>Data Col 2</TreeCol>
-                        <TreeCol>Data Col 3</TreeCol>
-                        <TreeCol>Data Col 4</TreeCol>
-                    </TreeRow>
-                    <TreeBranch>
-                        <TreeItem>
-                            <TreeRow>
-                                <TreeCol>Grandchild 1</TreeCol>
-                                <TreeCol>Data Col 2</TreeCol>
-                                <TreeCol>Data Col 3</TreeCol>
-                                <TreeCol>Data Col 4</TreeCol>
-                            </TreeRow>
-                            <TreeBranch>
-                                <TreeItem>
-                                    <TreeRow>
-                                        <TreeCol>Great Grandchild 1</TreeCol>
-                                        <TreeCol>Data Col 2</TreeCol>
-                                        <TreeCol>Data Col 3</TreeCol>
-                                        <TreeCol>Data Col 4</TreeCol>
-                                    </TreeRow>
-                                </TreeItem>
-                            </TreeBranch>
-                        </TreeItem>
-                    </TreeBranch>
-                </TreeItem>
-                <TreeItem>
-                    <TreeRow>
-                        <TreeCol>Child 2</TreeCol>
-                        <TreeCol>Data Col 2</TreeCol>
-                        <TreeCol>Data Col 3</TreeCol>
-                        <TreeCol>Data Col 4</TreeCol>
-                    </TreeRow>
-                </TreeItem>
-            </TreeBranch>
-        </TreeItem>
-        <TreeItem>
-            <TreeRow>
-                <TreeCol>Row 2</TreeCol>
-                <TreeCol>Data Col 2</TreeCol>
-                <TreeCol>Data Col 3</TreeCol>
-                <TreeCol>Data Col 4</TreeCol>
-            </TreeRow>
-            <TreeBranch>
-                <TreeItem>
-                    <TreeRow>
-                        <TreeCol>Child 1</TreeCol>
-                        <TreeCol>Data Col 2</TreeCol>
-                        <TreeCol>Data Col 3</TreeCol>
-                        <TreeCol>Data Col 4</TreeCol>
-                    </TreeRow>
-                </TreeItem>
-                <TreeItem>
-                    <TreeRow>
-                        <TreeCol>Child 2</TreeCol>
-                        <TreeCol>Data Col 2</TreeCol>
-                        <TreeCol>Data Col 3</TreeCol>
-                        <TreeCol>Data Col 4</TreeCol>
-                    </TreeRow>
-                </TreeItem>
-            </TreeBranch>
-        </TreeItem>
-        <TreeItem>
-            <TreeRow>
-                <TreeCol>Row 3</TreeCol>
-                <TreeCol>Data Col 2</TreeCol>
-                <TreeCol>Data Col 3</TreeCol>
-                <TreeCol>Data Col 4</TreeCol>
-            </TreeRow>
-        </TreeItem>
-        <TreeItem>
-            <TreeRow>
-                <TreeCol>Row 4</TreeCol>
-                <TreeCol>Data Col 2</TreeCol>
-                <TreeCol>Data Col 3</TreeCol>
-                <TreeCol>Data Col 4</TreeCol>
-            </TreeRow>
-            <TreeBranch>
-                <TreeItem>
-                    <TreeRow>
-                        <TreeCol>Child 1</TreeCol>
-                        <TreeCol>Data Col 2</TreeCol>
-                        <TreeCol>Data Col 3</TreeCol>
-                        <TreeCol>Data Col 4</TreeCol>
-                    </TreeRow>
-                </TreeItem>
-            </TreeBranch>
-        </TreeItem>
-    </Tree>
+    <TreeView.Head>
+        <TreeView.Col>Column Header 1</TreeView.Col>
+        <TreeView.Col>Column Header 2</TreeView.Col>
+        <TreeView.Col>Column Header 3</TreeView.Col>
+        <TreeView.Col>Column Header 4</TreeView.Col>
+    </TreeView.Head>
+    <TreeView.Tree>
+        <TreeView.Item>
+            <TreeView.Row>
+                <TreeView.Col>Row 1</TreeView.Col>
+                <TreeView.Col>Data Col 2</TreeView.Col>
+                <TreeView.Col>Data Col 3</TreeView.Col>
+                <TreeView.Col>Data Col 4</TreeView.Col>
+            </TreeView.Row>
+            <TreeView.Branch>
+                <TreeView.Item>
+                    <TreeView.Row>
+                        <TreeView.Col>Child 1</TreeView.Col>
+                        <TreeView.Col>Data Col 2</TreeView.Col>
+                        <TreeView.Col>Data Col 3</TreeView.Col>
+                        <TreeView.Col>Data Col 4</TreeView.Col>
+                    </TreeView.Row>
+                    <TreeView.Branch>
+                        <TreeView.Item>
+                            <TreeView.Row>
+                                <TreeView.Col>Grandchild 1</TreeView.Col>
+                                <TreeView.Col>Data Col 2</TreeView.Col>
+                                <TreeView.Col>Data Col 3</TreeView.Col>
+                                <TreeView.Col>Data Col 4</TreeView.Col>
+                            </TreeView.Row>
+                            <TreeView.Branch>
+                                <TreeView.Item>
+                                    <TreeView.Row>
+                                        <TreeView.Col>Great Grandchild 1</TreeView.Col>
+                                        <TreeView.Col>Data Col 2</TreeView.Col>
+                                        <TreeView.Col>Data Col 3</TreeView.Col>
+                                        <TreeView.Col>Data Col 4</TreeView.Col>
+                                    </TreeView.Row>
+                                </TreeView.Item>
+                            </TreeView.Branch>
+                        </TreeView.Item>
+                    </TreeView.Branch>
+                </TreeView.Item>
+                <TreeView.Item>
+                    <TreeView.Row>
+                        <TreeView.Col>Child 2</TreeView.Col>
+                        <TreeView.Col>Data Col 2</TreeView.Col>
+                        <TreeView.Col>Data Col 3</TreeView.Col>
+                        <TreeView.Col>Data Col 4</TreeView.Col>
+                    </TreeView.Row>
+                </TreeView.Item>
+            </TreeView.Branch>
+        </TreeView.Item>
+        <TreeView.Item>
+            <TreeView.Row>
+                <TreeView.Col>Row 2</TreeView.Col>
+                <TreeView.Col>Data Col 2</TreeView.Col>
+                <TreeView.Col>Data Col 3</TreeView.Col>
+                <TreeView.Col>Data Col 4</TreeView.Col>
+            </TreeView.Row>
+            <TreeView.Branch>
+                <TreeView.Item>
+                    <TreeView.Row>
+                        <TreeView.Col>Child 1</TreeView.Col>
+                        <TreeView.Col>Data Col 2</TreeView.Col>
+                        <TreeView.Col>Data Col 3</TreeView.Col>
+                        <TreeView.Col>Data Col 4</TreeView.Col>
+                    </TreeView.Row>
+                </TreeView.Item>
+                <TreeView.Item>
+                    <TreeView.Row>
+                        <TreeView.Col>Child 2</TreeView.Col>
+                        <TreeView.Col>Data Col 2</TreeView.Col>
+                        <TreeView.Col>Data Col 3</TreeView.Col>
+                        <TreeView.Col>Data Col 4</TreeView.Col>
+                    </TreeView.Row>
+                </TreeView.Item>
+            </TreeView.Branch>
+        </TreeView.Item>
+        <TreeView.Item>
+            <TreeView.Row>
+                <TreeView.Col>Row 3</TreeView.Col>
+                <TreeView.Col>Data Col 2</TreeView.Col>
+                <TreeView.Col>Data Col 3</TreeView.Col>
+                <TreeView.Col>Data Col 4</TreeView.Col>
+            </TreeView.Row>
+        </TreeView.Item>
+        <TreeView.Item>
+            <TreeView.Row>
+                <TreeView.Col>Row 4</TreeView.Col>
+                <TreeView.Col>Data Col 2</TreeView.Col>
+                <TreeView.Col>Data Col 3</TreeView.Col>
+                <TreeView.Col>Data Col 4</TreeView.Col>
+            </TreeView.Row>
+            <TreeView.Branch>
+                <TreeView.Item>
+                    <TreeView.Row>
+                        <TreeView.Col>Child 1</TreeView.Col>
+                        <TreeView.Col>Data Col 2</TreeView.Col>
+                        <TreeView.Col>Data Col 3</TreeView.Col>
+                        <TreeView.Col>Data Col 4</TreeView.Col>
+                    </TreeView.Row>
+                </TreeView.Item>
+            </TreeView.Branch>
+        </TreeView.Item>
+    </TreeView.Tree>
 </TreeView>`;
 
     const treeViewCodeRich = `
 <TreeView>
-    <TreeHead>
-        <TreeCol>Column Header 1</TreeCol>
-        <TreeCol>Column Header 2</TreeCol>
-        <TreeCol>Column Header 3</TreeCol>
-        <TreeCol>Column Header 4</TreeCol>
-    </TreeHead>
-    <Tree>
-        <TreeItem>
-            <TreeRow>
-                <TreeCol>Row 1</TreeCol>
-            </TreeRow>
-            <TreeBranch>
-                <TreeItem>
-                    <TreeRow>
-                        <TreeCol />
-                        <TreeCol><a href='http://www.google.com'>Google</a></TreeCol>
-                        <TreeCol><a href='http://www.bing.com'>Bing</a></TreeCol>
-                        <TreeCol><a href='http://www.yahoo.com'>Yahoo</a></TreeCol>
-                    </TreeRow>
-                </TreeItem>
-            </TreeBranch>
-        </TreeItem>
-    </Tree>
+    <TreeView.Head>
+        <TreeView.Col>Column Header 1</TreeView.Col>
+        <TreeView.Col>Column Header 2</TreeView.Col>
+        <TreeView.Col>Column Header 3</TreeView.Col>
+        <TreeView.Col>Column Header 4</TreeView.Col>
+    </TreeView.Head>
+    <TreeView.Tree>
+        <TreeView.Item>
+            <TreeView.Row>
+                <TreeView.Col>Row 1</TreeView.Col>
+            </TreeView.Row>
+            <TreeView.Branch>
+                <TreeView.Item>
+                    <TreeView.Row>
+                        <TreeView.Col />
+                        <TreeView.Col><a href='http://www.google.com'>Google</a></TreeView.Col>
+                        <TreeView.Col><a href='http://www.bing.com'>Bing</a></TreeView.Col>
+                        <TreeView.Col><a href='http://www.yahoo.com'>Yahoo</a></TreeView.Col>
+                    </TreeView.Row>
+                </TreeView.Item>
+            </TreeView.Branch>
+        </TreeView.Item>
+    </TreeView.Tree>
 </TreeView>`;
 
 
     return (
         <div>
-            <Header>TreeView</Header>
+            <Header>Tree View</Header>
             <Description>
-                A **TreeView** is used to display data in a visual hierarchy. Items that contain additional items
+                A **Tree View** is used to display data in a visual hierarchy. Items that contain additional items
                 are called nodes, while items that do not contain any other items are called leaves. If available,
                 a single top-most node is called a "root" node. Apart from the hierarchical structure of its nodes
                 and leaves, a tree is similar to a list.
@@ -217,54 +217,54 @@ export const TreeViewComponent = () => {
             </Description>
             <DocsTile>
                 <TreeView>
-                    <Tree>
-                        <TreeItem>
-                            <TreeRow>
-                                <TreeCol>Row 1</TreeCol>
-                            </TreeRow>
-                        </TreeItem>
-                        <TreeItem>
-                            <TreeRow>
-                                <TreeCol>Row 2</TreeCol>
-                            </TreeRow>
-                            <TreeBranch>
-                                <TreeItem>
-                                    <TreeRow>
-                                        <TreeCol>Child 1</TreeCol>
-                                    </TreeRow>
-                                </TreeItem>
-                                <TreeItem>
-                                    <TreeRow>
-                                        <TreeCol>Child 2</TreeCol>
-                                    </TreeRow>
-                                </TreeItem>
-                            </TreeBranch>
-                        </TreeItem>
-                        <TreeItem>
-                            <TreeRow>
-                                <TreeCol>Row 3</TreeCol>
-                            </TreeRow>
-                            <TreeBranch>
-                                <TreeItem>
-                                    <TreeRow>
-                                        <TreeCol>Child 1</TreeCol>
-                                    </TreeRow>
-                                    <TreeBranch>
-                                        <TreeItem>
-                                            <TreeRow>
-                                                <TreeCol>Grandchild 1</TreeCol>
-                                            </TreeRow>
-                                        </TreeItem>
-                                        <TreeItem>
-                                            <TreeRow>
-                                                <TreeCol>Grandchild 2</TreeCol>
-                                            </TreeRow>
-                                        </TreeItem>
-                                    </TreeBranch>
-                                </TreeItem>
-                            </TreeBranch>
-                        </TreeItem>
-                    </Tree>
+                    <TreeView.Tree>
+                        <TreeView.Item>
+                            <TreeView.Row>
+                                <TreeView.Col>Row 1</TreeView.Col>
+                            </TreeView.Row>
+                        </TreeView.Item>
+                        <TreeView.Item>
+                            <TreeView.Row>
+                                <TreeView.Col>Row 2</TreeView.Col>
+                            </TreeView.Row>
+                            <TreeView.Branch>
+                                <TreeView.Item>
+                                    <TreeView.Row>
+                                        <TreeView.Col>Child 1</TreeView.Col>
+                                    </TreeView.Row>
+                                </TreeView.Item>
+                                <TreeView.Item>
+                                    <TreeView.Row>
+                                        <TreeView.Col>Child 2</TreeView.Col>
+                                    </TreeView.Row>
+                                </TreeView.Item>
+                            </TreeView.Branch>
+                        </TreeView.Item>
+                        <TreeView.Item>
+                            <TreeView.Row>
+                                <TreeView.Col>Row 3</TreeView.Col>
+                            </TreeView.Row>
+                            <TreeView.Branch>
+                                <TreeView.Item>
+                                    <TreeView.Row>
+                                        <TreeView.Col>Child 1</TreeView.Col>
+                                    </TreeView.Row>
+                                    <TreeView.Branch>
+                                        <TreeView.Item>
+                                            <TreeView.Row>
+                                                <TreeView.Col>Grandchild 1</TreeView.Col>
+                                            </TreeView.Row>
+                                        </TreeView.Item>
+                                        <TreeView.Item>
+                                            <TreeView.Row>
+                                                <TreeView.Col>Grandchild 2</TreeView.Col>
+                                            </TreeView.Row>
+                                        </TreeView.Item>
+                                    </TreeView.Branch>
+                                </TreeView.Item>
+                            </TreeView.Branch>
+                        </TreeView.Item>
+                    </TreeView.Tree>
                 </TreeView>
             </DocsTile>
             <DocsText>{simpleTreeViewCode}</DocsText>
@@ -277,112 +277,112 @@ export const TreeViewComponent = () => {
             </Description>
             <DocsTile>
                 <TreeView>
-                    <TreeHead>
-                        <TreeCol>Column Header 1</TreeCol>
-                        <TreeCol>Column Header 2</TreeCol>
-                        <TreeCol>Column Header 3</TreeCol>
-                        <TreeCol>Column Header 4</TreeCol>
-                    </TreeHead>
-                    <Tree>
-                        <TreeItem>
-                            <TreeRow>
-                                <TreeCol>Row 1</TreeCol>
-                                <TreeCol>Data Col 2</TreeCol>
-                                <TreeCol>Data Col 3</TreeCol>
-                                <TreeCol>Data Col 4</TreeCol>
-                            </TreeRow>
-                            <TreeBranch>
-                                <TreeItem>
-                                    <TreeRow>
-                                        <TreeCol>Child 1</TreeCol>
-                                        <TreeCol>Data Col 2</TreeCol>
-                                        <TreeCol>Data Col 3</TreeCol>
-                                        <TreeCol>Data Col 4</TreeCol>
-                                    </TreeRow>
-                                    <TreeBranch>
-                                        <TreeItem>
-                                            <TreeRow>
-                                                <TreeCol>Grandchild 1</TreeCol>
-                                                <TreeCol>Data Col 2</TreeCol>
-                                                <TreeCol>Data Col 3</TreeCol>
-                                                <TreeCol>Data Col 4</TreeCol>
-                                            </TreeRow>
-                                            <TreeBranch>
-                                                <TreeItem>
-                                                    <TreeRow>
-                                                        <TreeCol>Great Grandchild 1</TreeCol>
-                                                        <TreeCol>Data Col 2</TreeCol>
-                                                        <TreeCol>Data Col 3</TreeCol>
-                                                        <TreeCol>Data Col 4</TreeCol>
-                                                    </TreeRow>
-                                                </TreeItem>
-                                            </TreeBranch>
-                                        </TreeItem>
-                                    </TreeBranch>
-                                </TreeItem>
-                                <TreeItem>
-                                    <TreeRow>
-                                        <TreeCol>Child 2</TreeCol>
-                                        <TreeCol>Data Col 2</TreeCol>
-                                        <TreeCol>Data Col 3</TreeCol>
-                                        <TreeCol>Data Col 4</TreeCol>
-                                    </TreeRow>
-                                </TreeItem>
-                            </TreeBranch>
-                        </TreeItem>
-                        <TreeItem>
-                            <TreeRow>
-                                <TreeCol>Row 2</TreeCol>
-                                <TreeCol>Data Col 2</TreeCol>
-                                <TreeCol>Data Col 3</TreeCol>
-                                <TreeCol>Data Col 4</TreeCol>
-                            </TreeRow>
-                            <TreeBranch>
-                                <TreeItem>
-                                    <TreeRow>
-                                        <TreeCol>Child 1</TreeCol>
-                                        <TreeCol>Data Col 2</TreeCol>
-                                        <TreeCol>Data Col 3</TreeCol>
-                                        <TreeCol>Data Col 4</TreeCol>
-                                    </TreeRow>
-                                </TreeItem>
-                                <TreeItem>
-                                    <TreeRow>
-                                        <TreeCol>Child 2</TreeCol>
-                                        <TreeCol>Data Col 2</TreeCol>
-                                        <TreeCol>Data Col 3</TreeCol>
-                                        <TreeCol>Data Col 4</TreeCol>
-                                    </TreeRow>
-                                </TreeItem>
-                            </TreeBranch>
-                        </TreeItem>
-                        <TreeItem>
-                            <TreeRow>
-                                <TreeCol>Row 3</TreeCol>
-                                <TreeCol>Data Col 2</TreeCol>
-                                <TreeCol>Data Col 3</TreeCol>
-                                <TreeCol>Data Col 4</TreeCol>
-                            </TreeRow>
-                        </TreeItem>
-                        <TreeItem>
-                            <TreeRow>
-                                <TreeCol>Row 4</TreeCol>
-                                <TreeCol>Data Col 2</TreeCol>
-                                <TreeCol>Data Col 3</TreeCol>
-                                <TreeCol>Data Col 4</TreeCol>
-                            </TreeRow>
-                            <TreeBranch>
-                                <TreeItem>
-                                    <TreeRow>
-                                        <TreeCol>Child 1</TreeCol>
-                                        <TreeCol>Data Col 2</TreeCol>
-                                        <TreeCol>Data Col 3</TreeCol>
-                                        <TreeCol>Data Col 4</TreeCol>
-                                    </TreeRow>
-                                </TreeItem>
-                            </TreeBranch>
-                        </TreeItem>
-                    </Tree>
+                    <TreeView.Head>
+                        <TreeView.Col>Column Header 1</TreeView.Col>
+                        <TreeView.Col>Column Header 2</TreeView.Col>
+                        <TreeView.Col>Column Header 3</TreeView.Col>
+                        <TreeView.Col>Column Header 4</TreeView.Col>
+                    </TreeView.Head>
+                    <TreeView.Tree>
+                        <TreeView.Item>
+                            <TreeView.Row>
+                                <TreeView.Col>Row 1</TreeView.Col>
+                                <TreeView.Col>Data Col 2</TreeView.Col>
+                                <TreeView.Col>Data Col 3</TreeView.Col>
+                                <TreeView.Col>Data Col 4</TreeView.Col>
+                            </TreeView.Row>
+                            <TreeView.Branch>
+                                <TreeView.Item>
+                                    <TreeView.Row>
+                                        <TreeView.Col>Child 1</TreeView.Col>
+                                        <TreeView.Col>Data Col 2</TreeView.Col>
+                                        <TreeView.Col>Data Col 3</TreeView.Col>
+                                        <TreeView.Col>Data Col 4</TreeView.Col>
+                                    </TreeView.Row>
+                                    <TreeView.Branch>
+                                        <TreeView.Item>
+                                            <TreeView.Row>
+                                                <TreeView.Col>Grandchild 1</TreeView.Col>
+                                                <TreeView.Col>Data Col 2</TreeView.Col>
+                                                <TreeView.Col>Data Col 3</TreeView.Col>
+                                                <TreeView.Col>Data Col 4</TreeView.Col>
+                                            </TreeView.Row>
+                                            <TreeView.Branch>
+                                                <TreeView.Item>
+                                                    <TreeView.Row>
+                                                        <TreeView.Col>Great Grandchild 1</TreeView.Col>
+                                                        <TreeView.Col>Data Col 2</TreeView.Col>
+                                                        <TreeView.Col>Data Col 3</TreeView.Col>
+                                                        <TreeView.Col>Data Col 4</TreeView.Col>
+                                                    </TreeView.Row>
+                                                </TreeView.Item>
+                                            </TreeView.Branch>
+                                        </TreeView.Item>
+                                    </TreeView.Branch>
+                                </TreeView.Item>
+                                <TreeView.Item>
+                                    <TreeView.Row>
+                                        <TreeView.Col>Child 2</TreeView.Col>
+                                        <TreeView.Col>Data Col 2</TreeView.Col>
+                                        <TreeView.Col>Data Col 3</TreeView.Col>
+                                        <TreeView.Col>Data Col 4</TreeView.Col>
+                                    </TreeView.Row>
+                                </TreeView.Item>
+                            </TreeView.Branch>
+                        </TreeView.Item>
+                        <TreeView.Item>
+                            <TreeView.Row>
+                                <TreeView.Col>Row 2</TreeView.Col>
+                                <TreeView.Col>Data Col 2</TreeView.Col>
+                                <TreeView.Col>Data Col 3</TreeView.Col>
+                                <TreeView.Col>Data Col 4</TreeView.Col>
+                            </TreeView.Row>
+                            <TreeView.Branch>
+                                <TreeView.Item>
+                                    <TreeView.Row>
+                                        <TreeView.Col>Child 1</TreeView.Col>
+                                        <TreeView.Col>Data Col 2</TreeView.Col>
+                                        <TreeView.Col>Data Col 3</TreeView.Col>
+                                        <TreeView.Col>Data Col 4</TreeView.Col>
+                                    </TreeView.Row>
+                                </TreeView.Item>
+                                <TreeView.Item>
+                                    <TreeView.Row>
+                                        <TreeView.Col>Child 2</TreeView.Col>
+                                        <TreeView.Col>Data Col 2</TreeView.Col>
+                                        <TreeView.Col>Data Col 3</TreeView.Col>
+                                        <TreeView.Col>Data Col 4</TreeView.Col>
+                                    </TreeView.Row>
+                                </TreeView.Item>
+                            </TreeView.Branch>
+                        </TreeView.Item>
+                        <TreeView.Item>
+                            <TreeView.Row>
+                                <TreeView.Col>Row 3</TreeView.Col>
+                                <TreeView.Col>Data Col 2</TreeView.Col>
+                                <TreeView.Col>Data Col 3</TreeView.Col>
+                                <TreeView.Col>Data Col 4</TreeView.Col>
+                            </TreeView.Row>
+                        </TreeView.Item>
+                        <TreeView.Item>
+                            <TreeView.Row>
+                                <TreeView.Col>Row 4</TreeView.Col>
+                                <TreeView.Col>Data Col 2</TreeView.Col>
+                                <TreeView.Col>Data Col 3</TreeView.Col>
+                                <TreeView.Col>Data Col 4</TreeView.Col>
+                            </TreeView.Row>
+                            <TreeView.Branch>
+                                <TreeView.Item>
+                                    <TreeView.Row>
+                                        <TreeView.Col>Child 1</TreeView.Col>
+                                        <TreeView.Col>Data Col 2</TreeView.Col>
+                                        <TreeView.Col>Data Col 3</TreeView.Col>
+                                        <TreeView.Col>Data Col 4</TreeView.Col>
+                                    </TreeView.Row>
+                                </TreeView.Item>
+                            </TreeView.Branch>
+                        </TreeView.Item>
+                    </TreeView.Tree>
                 </TreeView>
             </DocsTile>
             <DocsText>{treeViewCode}</DocsText>
@@ -395,29 +395,29 @@ export const TreeViewComponent = () => {
             </Description>
             <DocsTile>
                 <TreeView>
-                    <TreeHead>
-                        <TreeCol>Column Header 1</TreeCol>
-                        <TreeCol>Column Header 2</TreeCol>
-                        <TreeCol>Column Header 3</TreeCol>
-                        <TreeCol>Column Header 4</TreeCol>
-                    </TreeHead>
-                    <Tree>
-                        <TreeItem>
-                            <TreeRow>
-                                <TreeCol>Row 1</TreeCol>
-                            </TreeRow>
-                            <TreeBranch>
-                                <TreeItem>
-                                    <TreeRow>
-                                        <TreeCol />
-                                        <TreeCol><a href='http://www.google.com'>Google</a></TreeCol>
-                                        <TreeCol><a href='http://www.bing.com'>Bing</a></TreeCol>
-                                        <TreeCol><a href='http://www.yahoo.com'>Yahoo</a></TreeCol>
-                                    </TreeRow>
-                                </TreeItem>
-                            </TreeBranch>
-                        </TreeItem>
-                    </Tree>
+                    <TreeView.Head>
+                        <TreeView.Col>Column Header 1</TreeView.Col>
+                        <TreeView.Col>Column Header 2</TreeView.Col>
+                        <TreeView.Col>Column Header 3</TreeView.Col>
+                        <TreeView.Col>Column Header 4</TreeView.Col>
+                    </TreeView.Head>
+                    <TreeView.Tree>
+                        <TreeView.Item>
+                            <TreeView.Row>
+                                <TreeView.Col>Row 1</TreeView.Col>
+                            </TreeView.Row>
+                            <TreeView.Branch>
+                                <TreeView.Item>
+                                    <TreeView.Row>
+                                        <TreeView.Col />
+                                        <TreeView.Col><a href='http://www.google.com'>Google</a></TreeView.Col>
+                                        <TreeView.Col><a href='http://www.bing.com'>Bing</a></TreeView.Col>
+                                        <TreeView.Col><a href='http://www.yahoo.com'>Yahoo</a></TreeView.Col>
+                                    </TreeView.Row>
+                                </TreeView.Item>
+                            </TreeView.Branch>
+                        </TreeView.Item>
+                    </TreeView.Tree>
                 </TreeView>
             </DocsTile>
             <DocsText>{treeViewCodeRich}</DocsText>
